+++ Assets/Scripts/UX/UI/Ingame/ChrysosShop/ChrysosShop.cs
@@ -73,7 +73,7 @@ namespace UX.UI.Ingame.ChrysosShop
 
         public void Buy(PieceType type)
         {
-            MatchManager.Ins.InputProcessor.ExecuteAction(new ChrysosUpgrade(chrysos.Pos, new PieceConfig(type, chrysos.Color, candidate.Target), cost));
+            MatchManager.Ins.InputProcessor.ExecuteAction(new ChrysosUpgrade(chrysos.Pos, new PieceConfig(type, chrysos.Color, (ushort)candidate.Target), cost));
             Disable();
         }
     }
